Fix Int -> Long attributes

+++ trunk-vechain/src/main/scala/io/syspulse/haas/ingest/vechain/Transaction.scala
@@ -14,12 +14,12 @@ case class Transaction(
   nonce:String,     // nonce
   
   gas:BigInt,         // gas 
-  pric:Int,           // gas coefficient
+  pric:Long,           // gas coefficient
   
   data:String,      // calldata
 
   blk:String,       // blockref
-  exp:Int,          // expirateion
+  exp:Long,          // expirateion
   del:Option[String],       // delegator
   dep:Option[String],       // dependsOn
